Separate primitives into separate files
