+++ src/primitives/primitives.jl
@@ -1,0 +1,9 @@
+# each module is defined in a separate source file
+# each source file exports its own symbols
+include("flip.jl")
+include("uniform.jl")
+include("nil.jl")
+include("normal.jl")
+include("gamma.jl")
+include("mvnormal.jl")
+include("crp.jl")
